ISO Data Type v2.2 changes, which include:
NCI Localized 21090 ISO Data Types Library changes from v2.1.1 to v2.2

Affected data types:

AD
AD is primarily used to communicate data that will allow printing mail labels, or that will allow a person to physically visit that address. Addresses are essentially sequences of address parts, but add a "use" code and a valid time range for information about if and when the address can be used for a given purpose.
Add use : Set(PostalAddressUse) and the useablePeriod : QSET(TS).

CD
A CD is a reference to a concept defined in an external code system, terminology or ontology. In many cases, a CD is created from a value set – either a code/code system pair is chosen from a valueSet.  Translations represent a set of other CDs that each represent a translation of this CD into equivalent codes within the same code system or into corresponding concepts from other code systems.
Add valueSet : characterstring, valueSetVersion : characterstring, and translation : Set(CD).

ED
ED carries data that is primarily intended for human interpretation or for further machine processing outside the scope of this International Standard. This includes unformatted or formatted written language, multimedia data, or structured information as defined by a different standard (e.g., XML-signatures). The mediaType attribute identifies the type of the encapsulated data and can be used to determine a method to interpret or render the content. 
The IANA defined domain of media types is established by the IETF RFCs 2045 and 2046. mediaType has a default value of text/plain and cannot be null. If the media type is different to text/plain, the mediaType attribute shall be populated to be conformant in an HL7 communication.
Add mediaType : characterstring.

EN
An entity name is a name for a person, organization, place or thing.
Entity names are essentially sequences of entity name parts, but add a "use" code and a valid time range for information about when the name was used and how to choose between multiple aliases that may be valid at the same time.
Add use : Set(EntityNameUse).

QTY
QTY defines three facilities that all quantities may carry: an expression that may be used to derive the actual value, an originalText that carries the original form in which the quantity was represented and the uncertainty associated with the value.  These attributes are inherited for all data types that specialize QTY.
Add originalText : ED.TEXT and uncertainty : QTY.

PQ
PQ is a dimensioned quantity expressing the result of measuring. All PQ values must use UCUM codes for the units of measure. To accommodate the variance of expressions used, a translation attribute is used to provide an alternative representation of the same physical quantity expressed in a different unit from a different unit code system and possibly with a different value.
Add translation : Set(PQR).

SVN-Revision: 377